Use int64_t instead of long long

+++ number/combination.cpp
@@ -1,7 +1,7 @@
 #include "mod.cpp"
 
-vector<long long> fact;
-void init_fact(int n, long long mod) {
+vector<int64_t> fact;
+void init_fact(int n, int64_t mod) {
     fact.assign(n+1, 1);
     for (int i = 2; i <= n; ++i) {
         (fact[i] = fact[i-1] * i) %= mod;
@@ -9,6 +9,6 @@ void init_fact(int n, long long mod) {
 }
 
 // require init_fact(GREATER THAN OR EQUAL TO n, mod)
-long long C(long long n, long long r, long long mod) {
+int64_t C(int64_t n, int64_t r, int64_t mod) {
     return (fact[n] * invm((fact[r] * fact[n-r]) % mod, mod)) % mod;
 }
