+++ test/number/combination.cpp
@@ -3,14 +3,14 @@
 #include "../../number/combination.cpp"
 
 int main() {
-    long long n, k; cin >> n >> k;
-    const long long mod = 1e9+7;
+    int64_t n, k; cin >> n >> k;
+    const int64_t mod = 1e9+7;
 
     init_fact(k, mod);
 
-    long long ans = 0;
-    for (long long l = 0; l <= k; ++l) {
-        long long tmp = powm(k-l, n, mod);
+    int64_t ans = 0;
+    for (int64_t l = 0; l <= k; ++l) {
+        int64_t tmp = powm(k-l, n, mod);
         (tmp *= C(k, l, mod)) %= mod;
         if (l & 1) tmp = (-tmp + mod) % mod;
         (ans += tmp) %= mod;
